+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [2.1.5](https://github.com/bitrix24/b24ui/compare/v2.1.4...v2.1.5) (2025-11-xx)
+## [2.1.6](https://github.com/bitrix24/b24ui/compare/v2.1.5...v2.1.6) (2025-11-xx)
+
+* fix(Link): ensure consistency across Nuxt, Vue and Inertia
+
+## [2.1.5](https://github.com/bitrix24/b24ui/compare/v2.1.4...v2.1.5) (2025-11-19)
 
 ### Bug Fixes